Move fonts to separate module
